+++ great_expectations/dataset/sqlalchemy_dataset.py
@@ -248,8 +248,7 @@ class SqlAlchemyDataset(MetaSqlAlchemyDataset):
             #NOTE: Eugene 2020-01-31: @James, this is a not a proper fix, but without it the "public" schema
             #was used for a temp table and raising an error
             schema = None
-            # dashes are special characters in most databases so use underscores
-            table_name = "ge_tmp_" + str(uuid.uuid4()).replace("-", "_")
+            table_name = "ge_tmp_" + str(uuid.uuid4())[:8]
             generated_table_name = table_name
         else:
             generated_table_name = None
@@ -291,7 +290,11 @@ class SqlAlchemyDataset(MetaSqlAlchemyDataset):
         if schema is not None and custom_sql is not None:
             # temporary table will be written to temp schema, so don't allow
             # a user-defined schema
-            raise ValueError("Cannot specify both schema and custom_sql.")
+            # NOTE: 20200306 - JPC - Previously, this would disallow both custom_sql (a query) and a schema, but
+            # that is overly restrictive -- snowflake could have had a schema specified, for example, in which to create
+            # a temporary table.
+            # raise ValueError("Cannot specify both schema and custom_sql.")
+            pass
 
         if custom_sql is not None and self.engine.dialect.name.lower() == "bigquery":
             if generated_table_name is not None and self.engine.dialect.dataset_id is None:
@@ -300,11 +303,11 @@ class SqlAlchemyDataset(MetaSqlAlchemyDataset):
 
         if (custom_sql is not None and self.engine.dialect.name.lower() == "snowflake" and
                 generated_table_name is not None):
-            raise ValueError("No snowflake table_name specified. Snowflake with a query batch_kwarg will create "
+            raise ValueError("No snowflake_transient_table specified. Snowflake with a query batch_kwarg will create "
                              "a transient table, so you must provide a user-selected name.")
 
         if custom_sql:
-            self.create_temporary_table(table_name, custom_sql)
+            self.create_temporary_table(table_name, custom_sql, schema_name=schema)
 
             if generated_table_name is not None and self.engine.dialect.name.lower() == "bigquery":
                 logger.warning("Created permanent table {table_name}".format(
@@ -611,7 +614,7 @@ class SqlAlchemyDataset(MetaSqlAlchemyDataset):
 
         return self.engine.execute(query).scalar()
 
-    def create_temporary_table(self, table_name, custom_sql):
+    def create_temporary_table(self, table_name, custom_sql, schema_name=None):
         """
         Create Temporary table based on sql query. This will be used as a basis for executing expectations.
         WARNING: this feature is new in v0.4.
@@ -619,13 +622,40 @@ class SqlAlchemyDataset(MetaSqlAlchemyDataset):
         :param custom_sql:
         """
 
+        ###
+        # NOTE: 20200310 - The update to support snowflake transient table creation revealed several
+        # import cases that are not fully handled.
+        # The snowflake-related change updated behavior to allow both custom_sql and schema to be specified. But
+        # the underlying incomplete handling of schema remains.
+        #
+        # Several cases we need to consider:
+        #
+        # 1. Distributed backends (e.g. Snowflake and BigQuery) often use a `<database>.<schema>.<table>`
+        # syntax, but currently we are biased towards only allowing schema.table
+        #
+        # 2. In the wild, we see people using several ways to declare the schema they want to use:
+        # a. In the connection string, the original RFC only specifies database, but schema is supported by some
+        # backends (Snowflake) as a query parameter.
+        # b. As a default for a user (the equivalent of USE SCHEMA being provided at the beginning of a session)
+        # c. As part of individual queries.
+        #
+        # 3. We currently don't make it possible to select from a table in one query, but create a temporary table in
+        # another schema, except for with BigQuery and (now) snowflake, where you can specify the table name (and
+        # potentially trip of database, schema, table) in the batch_kwargs.
+        #
+        # The SqlAlchemyDataset interface essentially predates the batch_kwargs concept and so part of what's going
+        # on, I think, is a mismatch between those. I think we should rename custom_sql -> "temp_table_query" or
+        # similar, for example.
+        ###
+
         if self.engine.dialect.name.lower() == "bigquery":
-            logger.info("Creating table %s" % table_name)
             stmt = "CREATE OR REPLACE TABLE `{table_name}` AS {custom_sql}".format(
                 table_name=table_name, custom_sql=custom_sql)
         elif self.engine.dialect.name.lower() == "snowflake":
             logger.info("Creating transient table %s" % table_name)
-            stmt = "CREATE TRANSIENT TABLE {table_name} AS {custom_sql}".format(
+            if schema_name is not None:
+                table_name = schema_name + "." + table_name
+            stmt = "CREATE OR REPLACE TRANSIENT TABLE {table_name} AS {custom_sql}".format(
                 table_name=table_name, custom_sql=custom_sql)
         elif self.engine.dialect.name == "mysql":
             stmt = "CREATE TEMPORARY TABLE {table_name} AS {custom_sql}".format(